Store arg state in a named tuple instead of using exceptions
This makes it more explicit that an argument to open() not being present
is something that we expect as part of the normal execution of this
plugin. It also saves Python from having to generate the stack trace
objects, so it may have better performance (I haven't measured this,
however).